requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
azure-common==1.1.28
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
base58==2.1.1
bitarray==2.6.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
cytoolz==0.12.0
Django==3.2.16
django-storages==1.11.1
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.3.3
gunicorn==20.1.0
hexbytes==0.3.0
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.10.1
ipfshttpclient==0.8.0a2
jsonschema==4.17.3
lru-dict==1.1.8
multiaddr==0.0.9
multidict==6.0.3
netaddr==0.8.0
parsimonious==0.8.1
pkgutil_resolve_name==1.3.10
protobuf==3.19.5
pycparser==2.21
pycryptodome==3.16.0
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
rlp==2.0.1
six==1.16.0
sqlparse==0.4.3
toolz==0.12.0
typing_extensions==4.4.0
urllib3==1.26.13
varint==1.0.2
web3==5.31.3
websockets==9.1
yarl==1.8.2
zipp==3.11.0
